% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggRaster.R
\name{aggRaster}
\alias{aggRaster}
\title{aggRaster}
\usage{
aggRaster(x, func = "mean", inmem = FALSE, varnam = "wn",
  outdir = getwd())
}
\arguments{
\item{x, ind.months=, ouputdir}{}
}
\description{
Aggregate raster bricks from daily to monthly in parallel
}
\examples{
splash.grid()	
}
\keyword{splash}
