% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipPointGrid.R
\name{clipPointGrid}
\alias{clipPointGrid}
\title{clipPointGrid}
\usage{
clipPointGrid(raster, sp, value)
}
\arguments{
\item{raster, sp, value}{}
}
\description{
function to clip points(sp) in overlapped ares where raster == value
}
\examples{
splash.grid()
}
\keyword{splash}
