% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud2Rad.R
\name{cloud2Rad}
\alias{cloud2Rad}
\title{cloud2Rad}
\usage{
cloud2Rad(clds, elev_lowres, ouputdir = getwd(), inmem = FALSE, ...)
}
\arguments{
\item{elev_hres, rad_lowres, ouputdir}{}
}
\description{
Dowsncale solar radiation grid data by getting the transmitance from extraterrestrial radiation, and applying slope corrections, returns w/m2
}
\examples{
splash.grid()
}
\keyword{splash}
