% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleKrige.R
\name{downscaleKrige}
\alias{downscaleKrige}
\title{downscaleKrige}
\usage{
downscaleKrige(low_res_raster, high_res_raster, ouputdir = getwd())
}
\arguments{
\item{ouputdir, }{directory where to save the files}

\item{elev_hres, }{elevation, a high resolution Raster layer object}

\item{rad_lowres, }{solar radiation (w/m2), the low resolution Raster* object with z time dimension}
}
\value{
solar radiation (w/m2), Raster* object with the resolution of elev_hres and the same z time dimension as rad_lowres
}
\description{
Downscale the input applying the autokrige from hydroTSM::hydrokrige, using elevation as predictor
}
\examples{
# *optional run beginCluster() first, for parallel computing
downscaleKrige()
}
\keyword{Downscale,}
\keyword{radiation,}
\keyword{solar}
\keyword{terrain}
