% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleRLM.R
\name{downscaleRLM}
\alias{downscaleRLM}
\title{downscaleRLM}
\usage{
downscaleRLM(low_res_raster, high_res_raster)
}
\arguments{
\item{low_res_raster, }{raster layer, low resolution}

\item{high_res_raster, }{Raster* object}
}
\value{
Raster* object, with the same resolution as high_res_raster
}
\description{
Downscale precipitation or temperature using robust linear models, wrapper from the spatialEco package
}
\examples{
downscaleRLM()
}
\keyword{downscale}
