% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleSolar.R
\name{downscaleSolar}
\alias{downscaleSolar}
\title{downscaleSolar}
\usage{
downscaleSolar(elev_hres, rad_lowres, ouputdir)
}
\arguments{
\item{elev_hres, rad_lowres, ouputdir}{}
}
\description{
Dowsncale solar radiation grid data by getting the transmitance from extraterrestrial radiation, and applying slope corrections, returns w/m2
}
\examples{
splash.grid()
}
\keyword{splash}
