% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForcingNRCS.R
\name{getForcingNRCS}
\alias{getForcingNRCS}
\title{getForcingNRCS}
\usage{
getForcingNRCS(site_id, lat, lon)
}
\arguments{
\item{fields}{on stations.sp: lat, lon, code}
}
\description{
Wraper of the RNRCS and daymetr packages to get forcing and soil moisture data at NRCS sites
}
\examples{
splash.grid()
}
\keyword{splashTools}
