% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGLMchelsa.R
\name{getGLMchelsa}
\alias{getGLMchelsa}
\title{getGLMchelsa}
\usage{
getGLMchelsa(model, scenario, yr, outdir = getwd())
}
\arguments{
\item{model:}{character describing the glm, check http://chelsa-climate.org/downloads/}

\item{scenario:}{character: 'rcp45', 'rcp60' or 'rcp85'}

\item{yr:}{integer year of the projection, only available: 2060 or 2080}
}
\value{
list with monthly temperature and precipitation from the GLM's projections, both saved as netCDF files
}
\description{
Function to ease the download of the high resolution climate projections from the chelsa server
}
\examples{
getGLMchelsa(model='bcc',scenario='rcp85',yr=2060)
}
\keyword{splash}
