% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModisClim.R
\name{getModisClim}
\alias{getModisClim}
\title{getModisClim}
\usage{
getModisClim(lat, lon, start, end, outmode = list(tile = TRUE, monthly =
  TRUE, use.clouds = FALSE), dem, outdir = getwd(), usr = "usr",
  pass = "pass")
}
\arguments{
\item{start, }{end : data range}

\item{coords}{vector c(lat,lon)}
}
\description{
download and gapfill MOD07 and MYD07 atmospheric profiles, calculate saturated vapour pressure using dowsncaled LST from Microwave SSMI and Modis IR, if used with the option use.clouds=TRUE, additional files from MOD06 and MYD06 cloud product will be used to infer temperature and actual vapour pressure under the clouds below the tropopause.
}
\examples{
getModisClim()
}
\keyword{modis}
