% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoilNRCS.R
\name{getSoilNRCS}
\alias{getSoilNRCS}
\title{getSoilNRCS}
\usage{
getSoilNRCS(coords, maxdepth = NULL)
}
\arguments{
\item{coords}{vector c(lat,lon)}

\item{maxdepth}{maximum depth in  cm, to calculate the weigthed average, if maxdepth=NULL (default) the function uses all the available depths}
}
\description{
wrapper of the soilDB to retrievesoil data (sand, clay, OM, gravel, bulk density and depth) from the NRCS sites
}
\examples{
splash.grid()
}
\keyword{splash}
