% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoilSite.R
\name{getSoilSite}
\alias{getSoilSite}
\title{getSoilSite}
\usage{
getSoilSite(lat, lon)
}
\arguments{
\item{lat, }{lon}
}
\value{
a numeric vector with percentages of sand (w/w), clay(w/w), SOM(w/w), gravel(v/v); bulk density (g/cm3) and depth (m)
}
\description{
Get soil data from soilgrids REST API, point data is from the 250m resolution dataset, calculates the weigthed mean according layer depths
}
\examples{
getSoilSite(lat=8.7333,lon=-70.883333) 
}
\keyword{soil}
\keyword{texture}
