% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateForcing.R
\name{interpolateForcing}
\alias{interpolateForcing}
\title{interpolateForcing}
\usage{
interpolateForcing(stations.sp, data.df, dem, outdir = getwd())
}
\arguments{
\item{fields}{on stations.sp: lat, lon, code}
}
\description{
Calculates energy and water fluxes using splash routines, spins up until reaching steady state, then do the caldulations
}
\examples{
splash.grid()
}
\keyword{splash}
