% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krigeForcing.R
\name{krigeForcing}
\alias{krigeForcing}
\title{krigeForcing}
\usage{
krigeForcing(stations.sp, data.df, dem, outdir = getwd())
}
\arguments{
\item{stations.sp:}{SpatialPointsDataFrame object, mandatory attributes : lat, lon, code of the station, min 4 points}

\item{data.df:}{xts matrix, names of the columns should be the same as in stations.sp$code}

\item{dem:}{Raster layer, elevation (masl)}
}
\value{
A RasterBrick object with z time dimension saved to outdir as a netcdf file
}
\description{
Wrapper of the hydroTSM::hydrokrige to apply automatic kriging over time series in parallel*
}
\examples{
*optional run beginCluster() first, for parallel computing
krigeForcing()
}
\keyword{splash}
