% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetForcing.R
\name{subsetForcing}
\alias{subsetForcing}
\title{subsetForcing}
\usage{
subsetForcing(forcing_path, varnam, dem, dem_lowres, yr.start, yr.end,
  ouputdir = getwd(), downscale = TRUE)
}
\arguments{
\item{forcing_path, varnam, dem, dem_lowres, yr.start, yr.end, ouputdir}{}
}
\description{
Subset spatially and temporally low resolution data (currently only WFDEI) and do the downscaling
}
\examples{
splash.grid()
}
\keyword{splash}
