% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsSWC.grid.R
\name{unSWC.grid}
\alias{unSWC.grid}
\title{unsSWC.grid}
\usage{
unSWC.grid(soil_data, uns_depth, wn)
}
\arguments{
\item{soil_data, }{vector retrieved from getSoilNRCS() or getSoilSite()}

\item{uns_depth, }{depth to where to calculate water content in (m)}

\item{wn, }{xts object, total water content of the profile from surface to bedrock (m)}

\item{ouputdir, }{directory where to save the files}
}
\description{
Estimates water content using an analytca integral of the profile described by Brooks and Corey 1964 and calculates depth to water table if there is any
}
\examples{
unSWC()
}
\keyword{splashTools}
