% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggRaster.R
\name{aggRaster}
\alias{aggRaster}
\title{aggRaster}
\usage{
aggRaster(x, func = "mean", inmem = FALSE, outdir = getwd(), ...)
}
\arguments{
\item{x, }{Raster* object}

\item{func, }{character, 'mean' or ,'sum'}

\item{inmem, }{logical, if FALSE the result will be stored as a netCDF file}

\item{outdir, }{character, directory path where the output will be stored if inmem=FALSE}

\item{...}{additional arguments as for writeRaster, must include varnam,longname, an varunit}
}
\value{
Rasterbrick object with monthly z time dimension
}
\description{
Aggregate raster bricks from daily to monthly in parallel
}
\examples{
*optional run beginCluster() first, for parallel computing
aggRaster()	
}
\keyword{aggregation,}
\keyword{monthly}
