% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipGrid.R
\name{clipGrid}
\alias{clipGrid}
\title{clipGrid}
\usage{
clipGrid(raster, poly)
}
\arguments{
\item{raster, }{Raster* object}

\item{poly, }{a SpatialPolygons* object}
}
\value{
A Raster* object
}
\description{
clips and mask a raster object, based on a polygon
}
\examples{
clipGrid()
}
\keyword{splash}
