% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipPointGrid.R
\name{clipPointGrid}
\alias{clipPointGrid}
\title{clipPointGrid}
\usage{
clipPointGrid(raster, sp, value)
}
\arguments{
\item{raster, }{Raster layer}

\item{sp, }{SpatialPoints or SpatialPointsDataFrame object}

\item{value, }{numeric}
}
\value{
A subset of sp
}
\description{
function to clip points(sp) in overlapped ares where raster == value
}
\examples{
clipPointGrid()
}
\keyword{splash}
