% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud2Rad.R
\name{cloud2Rad}
\alias{cloud2Rad}
\title{cloud2Rad
Compute mean solar radiation (W/m2 ) by using the classic AP formula and the elevation effect on the atmosphere's transmittance, as described by Davis et al., (2017)}
\usage{
cloud2Rad(clds, elev_lowres, ouputdir = getwd(), inmem = FALSE, ...)
}
\arguments{
\item{clds, }{cloudiness, (fraction), Raster* object with z time dimension}

\item{elev_lowres, }{elevation (masl), Raster layer}

\item{ouputdir, }{directory where to save the files}
}
\value{
A RasterBrick object with the mean solar radiation (W/m2 ) with the same dimensions as clds
}
\description{
cloud2Rad
Compute mean solar radiation (W/m2 ) by using the classic AP formula and the elevation effect on the atmosphere's transmittance, as described by Davis et al., (2017)
}
\examples{
*optional run beginCluster() first, for parallel computing
cloud2Rad()
}
\keyword{splash}
