% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleRLM.R
\name{downscaleRLM}
\alias{downscaleRLM}
\title{downscaleRLM}
\usage{
downscaleRLM(low_res_raster, high_res_raster)
}
\arguments{
\item{fields}{on stations.sp: lat, lon, code}
}
\description{
Downscale precipitation or temperature using robust linear models, wrapper from the spatialEco package
}
\examples{
splash.grid()
}
\keyword{splashTools}
