% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModisClim.R
\name{getModisClim}
\alias{getModisClim}
\title{getModisClim}
\usage{
getModisClim(lat, lon, start, end, outmode = list(tile = TRUE, monthly =
  TRUE, use.clouds = TRUE), dem, outdir = getwd(), usr = "usr",
  pass = "pass")
}
\arguments{
\item{start, }{end : data range}

\item{coords}{vector c(lat,lon)}
}
\description{
download and gapfill MOD06 and MYD06 atmospheric profiles, calculate es sing dowsncaled LST from Microwave and Modis IR
}
\examples{
getModisClim()
}
\keyword{splash}
