% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoilGrid.R
\name{getSoilGrid}
\alias{getSoilGrid}
\title{getSoilGrid}
\usage{
getSoilGrid(bbox, ouputdir = getwd())
}
\arguments{
\item{bbox, }{any type of spatial object}

\item{ouputdir, }{directory where to save the files}
}
\description{
Downloads soil data (sand, clay, OM, gravel, bulk density and depth) from soilgrids.org in raster format, calculates the weigthed mean according layer depths
}
\examples{
splash.grid()
}
\keyword{splashTools}
