% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSoilSite.R
\name{getSoilSite}
\alias{getSoilSite}
\title{getSoilSite}
\usage{
getSoilSite(lat, lon)
}
\arguments{
\item{lat, }{lon}
}
\description{
Get soil data (sand, clay, OM, gravel, bulk density and depth) from soilgrids REST API, point data is from the 250m resolution dataset, calculates the weigthed mean according layer depths
}
\examples{
splash.grid()
}
\keyword{splash}
