% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateForcing.R
\name{interpolateForcing}
\alias{interpolateForcing}
\title{interpolateForcing}
\usage{
interpolateForcing(stations.sp, data.df, dem, outdir = getwd())
}
\arguments{
\item{fields}{on stations.sp: lat, lon, code}
}
\description{
Wrapper of the hydroTSM::hydrokrige to interpolate over time series in parallel
}
\examples{
splash.grid()
}
\keyword{splash}
